na classe rest *, +, -

+++ src/main/java/br/com/db1/dao/CalculadoraSimplesDao.java
@@ -2,19 +2,19 @@ package br.com.db1.dao;
 
 public class CalculadoraSimplesDao {
 	
-	public Integer dividir(Integer valor1, Integer valor2) {
+	public Double dividir(Double valor1, Double valor2) {
 		return valor1 / valor2;
 	}
 	
-	public Integer multiplicar(Integer valor1, Integer valor2) {
+	public Double multiplicar(Double valor1, Double valor2) {
 		return valor1 * valor2;
 	}
 	
-	public Integer somar(Integer valor1, Integer valor2) {
+	public Double somar(Double valor1, Double valor2) {
 		return valor1 + valor2;
 	}
 	
-	public Integer subtrair(Integer valor1, Integer valor2) {
+	public Double subtrair(Double valor1, Double valor2) {
 		return valor1 - valor2;
 	}
 }
